# Optimize arrayReduce, -Array and -State combinators