--- a/dbms/tests/performance/vectorize_aggregation_combinators.xml
+++ b/dbms/tests/performance/vectorize_aggregation_combinators.xml
@@ -1,0 +1,35 @@
+<test>
+
+    <type>loop</type>
+
+    <stop_conditions>
+        <all_of>
+            <total_time_ms>30000</total_time_ms>
+        </all_of>
+        <any_of>
+            <average_speed_not_changing_for_ms>6000</average_speed_not_changing_for_ms>
+            <total_time_ms>60000</total_time_ms>
+        </any_of>
+    </stop_conditions>
+
+    <main_metric>
+        <min_time/>
+    </main_metric>
+
+    <settings>
+        <max_threads>1</max_threads>
+    </settings>
+
+    <create_query>CREATE TABLE array_data(k UInt16, v Array(UInt64)) ENGINE Log</create_query>
+
+    <fill_query>INSERT INTO array_data SELECT number % 1024, arrayWithConstant(16, number) from numbers(10000000)</fill_query>
+
+    <query>SELECT countMerge(v) FROM (SELECT countState() v FROM numbers(1000000000)) FORMAT Null</query>
+    <query>SELECT countMerge(v) FROM (SELECT number % 1024 k, countState() v FROM numbers(1000000000) GROUP BY k) FORMAT Null</query>
+
+    <query>SELECT sumArray(v) FROM array_data FORMAT Null</query>
+    <query>SELECT k, sumArray(v) FROM array_data GROUP BY k FORMAT Null</query>
+    <query>SELECT arrayReduce('avg', v) FROM array_data FORMAT Null</query>
+
+    <drop_query>DROP TABLE IF EXISTS array_data</drop_query>
+</test>