I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

For changelog. Remove if this is non-significant change.

Category (leave one):
- Performance Improvement


Short description (up to few sentences):
vectorize processing arrayReduce similar to Aggregator addBatch. Might also serve as an infrastructure for https://github.com/ClickHouse/ClickHouse/pull/7550


```
native:
201.nobida.cn :) select ignore(arraySum(a)) from aa format Null

SELECT ignore(arraySum(a))
FROM aa
FORMAT Null


10000000 rows in set. Elapsed: 0.423 sec. Processed 10.00 million rows, 880.00 MB (23.66 million rows/s., 2.08 GB/s.)


before:
201.nobida.cn :) select ignore(arrayReduce('sum', a)) from aa format Null

SELECT ignore(arrayReduce('sum', a))
FROM aa
FORMAT Null


10000000 rows in set. Elapsed: 0.957 sec. Processed 10.00 million rows, 880.00 MB (10.45 million rows/s., 919.44 MB/s.)



after:
201.nobida.cn :) select ignore(arrayReduce('sum', a)) from aa format Null

SELECT ignore(arrayReduce('sum', a))
FROM aa
FORMAT Null


10000000 rows in set. Elapsed: 0.583 sec. Processed 10.00 million rows, 880.00 MB (17.17 million rows/s., 1.51 GB/s.)
```