@alexey-milovidov 2019-11-04
Results of performance test look strange: https://clickhouse-test-reports.s3.yandex.net/7608/5c160d5efb0ca08b6b515d66d2abb0bb847713b4/performance_test/comparison_min_time_gcc_9.html

Could you please compare the most different result manually with `clickhouse-benchmark`?