@amosbird 2019-11-05
# Summary

This pr optimizes three use cases.

1. arrayReduce:  0.9 sec -> 0.56 sec, **60%** improvement
2. -Array combinator: 0.9 sec -> 0.34 sec, **260%** improvement
3. -State combinator: 3.0 sec -> 0.5 sec,  **600%** improvement

The actual improvements depend on aggregate functions, group by methods, group by size, etc.

Supercede https://github.com/ClickHouse/ClickHouse/pull/7508  https://github.com/ClickHouse/ClickHouse/pull/7549

# clickhoues-benchmark results

- Query: select arrayReduce('sum', a) from aa format Null;

```
Before:

Queries executed: 51.

localhost:9000, queries 51, QPS: 1.014, RPS: 10142086.101, MiB/s: 851.158, result RPS: 10142086.101, result MiB/s: 77.378.

0.000%          0.908 sec.
10.000%         0.938 sec.
20.000%         0.942 sec.
30.000%         0.944 sec.
40.000%         0.948 sec.
50.000%         0.951 sec.
60.000%         0.956 sec.
70.000%         0.961 sec.
80.000%         1.077 sec.
90.000%         1.093 sec.
95.000%         1.142 sec.
99.000%         1.194 sec.
99.900%         1.234 sec.
99.990%         1.238 sec.

After:

Queries executed: 59.

localhost:9000, queries 59, QPS: 1.681, RPS: 16814842.041, MiB/s: 1411.158, result RPS: 16814842.041, result MiB/s: 128.287.

0.000%          0.560 sec.
10.000%         0.582 sec.
20.000%         0.587 sec.
30.000%         0.589 sec.
40.000%         0.592 sec.
50.000%         0.595 sec.
60.000%         0.597 sec.
70.000%         0.602 sec.
80.000%         0.604 sec.
90.000%         0.607 sec.
95.000%         0.610 sec.
99.000%         0.632 sec.
99.900%         0.635 sec.
99.990%         0.636 sec.
```

- Query: select sumArray(a) from aa format Null;

```
Before:

Queries executed: 51.

localhost:9000, queries 51, QPS: 1.045, RPS: 10453042.085, MiB/s: 877.254, result RPS: 1.045, result MiB/s: 0.000.

0.000%          0.897 sec.
10.000%         0.928 sec.
20.000%         0.931 sec.
30.000%         0.937 sec.
40.000%         0.949 sec.
50.000%         0.951 sec.
60.000%         0.954 sec.
70.000%         0.968 sec.
80.000%         0.978 sec.
90.000%         0.990 sec.
95.000%         0.994 sec.
99.000%         1.065 sec.
99.900%         1.078 sec.
99.990%         1.079 sec.

After:

Queries executed: 54.

localhost:9000, queries 54, QPS: 2.642, RPS: 26421096.796, MiB/s: 2217.347, result RPS: 2.642, result MiB/s: 0.000.

0.000%          0.345 sec.
10.000%         0.375 sec.
20.000%         0.377 sec.
30.000%         0.378 sec.
40.000%         0.379 sec.
50.000%         0.380 sec.
60.000%         0.381 sec.
70.000%         0.381 sec.
80.000%         0.382 sec.
90.000%         0.384 sec.
95.000%         0.385 sec.
99.000%         0.389 sec.
99.900%         0.390 sec.
99.990%         0.390 sec.
```

- Query: select countState() from numbers(1000000000) format Null;

```
Before:

Queries executed: 21.

localhost:9000, queries 21, QPS: 0.324, RPS: 324260324.149, MiB/s: 2473.910, result RPS: 0.324, result MiB/s: 0.000.

0.000%          3.020 sec.
10.000%         3.041 sec.
20.000%         3.047 sec.
30.000%         3.049 sec.
40.000%         3.052 sec.
50.000%         3.054 sec.
60.000%         3.056 sec.
70.000%         3.060 sec.
80.000%         3.083 sec.
90.000%         3.100 sec.
95.000%         3.130 sec.
99.000%         3.481 sec.
99.900%         3.560 sec.
99.990%         3.568 sec.


After:

Queries executed: 41.

localhost:9000, queries 41, QPS: 1.896, RPS: 1895701598.912, MiB/s: 14463.055, result RPS: 1.896, result MiB/s: 0.000.

0.000%          0.490 sec.
10.000%         0.512 sec.
20.000%         0.516 sec.
30.000%         0.521 sec.
40.000%         0.522 sec.
50.000%         0.526 sec.
60.000%         0.530 sec.
70.000%         0.538 sec.
80.000%         0.540 sec.
90.000%         0.544 sec.
95.000%         0.549 sec.
99.000%         0.571 sec.
99.900%         0.573 sec.
99.990%         0.574 sec.
```