@amosbird 2019-11-05
The perf results don't make any sense to me. The slow down queries aren't related to any of this pr's changes