@amosbird 2019-11-05
btw, some perf tests use `total_time` as its main_metric, which is confusing.