@alexey-milovidov 2019-11-05
I have checked manually the query that has the most slowdown on performance test:

```
SELECT SearchEngineID AS k, uniqCombined(18)(UserID) FROM hits_100m_single GROUP BY k
```

I have downloaded two binaries, one from this PR and one from master:
```
wget https://clickhouse-builds.s3.yandex.net/0/6a871f579fa9638acc08ac8275f3aa1d2736a8f6/1572912744_deb/clickhouse-common-static_19.17.1.1613_amd64.deb
wget https://clickhouse-builds.s3.yandex.net/7608/206d959d1f8bc9c61aa8c90608862ba89cbd6852/1572933436_deb/clickhouse-common-static_19.17.1.1614_amd64.deb
```

Unpacked them to `clickhouse-19.17-old` and `clickhouse-19.17-new`.
Run them from the terminal:
```
sudo -u clickhouse ./clickhouse-19.17-old server -C /etc/clickhouse-server/config.xml
```

And run clickhouse-benchmark:

```
Old:

Queries executed: 54.

localhost:9000, queries 54, QPS: 2.751, RPS: 275118567.607, MiB/s: 2623.735, result RPS: 264.114, result MiB/s: 0.003.

0.000%          0.314 sec.
10.000%         0.332 sec.
20.000%         0.340 sec.
30.000%         0.344 sec.
40.000%         0.349 sec.
50.000%         0.357 sec.
60.000%         0.368 sec.
70.000%         0.372 sec.
80.000%         0.393 sec.
90.000%         0.407 sec.
95.000%         0.413 sec.
99.000%         0.422 sec.
99.900%         0.425 sec.
99.990%         0.426 sec.

New:

Queries executed: 57.

localhost:9000, queries 57, QPS: 2.603, RPS: 260334492.752, MiB/s: 2482.743, result RPS: 249.921, result MiB/s: 0.002.

0.000%          0.344 sec.
10.000%         0.352 sec.
20.000%         0.356 sec.
30.000%         0.367 sec.
40.000%         0.374 sec.
50.000%         0.382 sec.
60.000%         0.386 sec.
70.000%         0.398 sec.
80.000%         0.410 sec.
90.000%         0.418 sec.
95.000%         0.439 sec.
99.000%         0.448 sec.
99.900%         0.450 sec.
99.990%         0.450 sec.


Old:

Queries executed: 211.

localhost:9000, queries 211, QPS: 3.020, RPS: 302000223.497, MiB/s: 2880.099, result RPS: 289.920, result MiB/s: 0.003.

0.000%          0.290 sec.
10.000%         0.299 sec.
20.000%         0.307 sec.
30.000%         0.310 sec.
40.000%         0.318 sec.
50.000%         0.324 sec.
60.000%         0.329 sec.
70.000%         0.340 sec.
80.000%         0.359 sec.
90.000%         0.377 sec.
95.000%         0.392 sec.
99.000%         0.404 sec.
99.900%         0.410 sec.
99.990%         0.411 sec.

New:

Queries executed: 224.

localhost:9000, queries 224, QPS: 2.607, RPS: 260723585.248, MiB/s: 2486.454, result RPS: 250.295, result MiB/s: 0.002.

0.000%          0.336 sec.
10.000%         0.350 sec.
20.000%         0.357 sec.
30.000%         0.361 sec.
40.000%         0.368 sec.
50.000%         0.375 sec.
60.000%         0.384 sec.
70.000%         0.395 sec.
80.000%         0.411 sec.
90.000%         0.436 sec.
95.000%         0.446 sec.
99.000%         0.458 sec.
99.900%         0.471 sec.
99.990%         0.472 sec.
```

And have found that this performance regression is real.

Although the code looks reasonable and performance improvement that you've demonstrated is massive.

Let's look at perf top.