@amosbird 2019-11-06
Ok, it seems some compiler tweaks are required. I'll investigate...

> This is the new feature of clickhouse-benchmark that should guarantee statistical significant results

That's awesome!